--- a/src/model.ts
+++ b/src/model.ts
@@ -381,13 +381,15 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
       if (operations.length === 0) {
         return;
       }
-      const finalOperations = operations.map((op) => {
+
+      const finalOperations = [];
+      for (const op of operations) {
         let operation = op;
         if ('insertOne' in op) {
           operation = {
             insertOne: {
               document: {
-                ...getDefaultValues(model.defaults),
+                ...(await getDefaultValues(model.defaults)),
                 ...op.insertOne.document,
               },
             },
@@ -405,7 +407,7 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
                 ...update,
                 $setOnInsert: cleanSetOnInsert(
                   {
-                    ...getDefaultValues(model.defaults),
+                    ...(await getDefaultValues(model.defaults)),
                     ...update.$setOnInsert,
                   },
                   update
@@ -414,10 +416,11 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
             },
           };
         }
-        return model.timestamps
-          ? timestampBulkWriteOperation(operation, model.timestamps)
-          : operation;
-      });
+
+        finalOperations.push(
+          model.timestamps ? timestampBulkWriteOperation(operation, model.timestamps) : operation
+        );
+      }
 
       const result = await model.collection.bulkWrite(
         finalOperations as AnyBulkWriteOperation<TSchema>[],
@@ -826,7 +829,7 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
     };
 
     const $setOnInsert = cleanSetOnInsert({
-      ...getDefaultValues(model.defaults),
+      ...await getDefaultValues(model.defaults),
       ...finalUpdate.$setOnInsert,
       ...created,
     }, finalUpdate);
@@ -868,16 +871,17 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
       docs: DocumentForInsert<TSchema, TOptions>[],
       options?: BulkWriteOptions
     ): Promise<TSchema[]> {
-      const documents = docs.map((doc) => {
-        return {
+      const documents = [];
+      for (const doc of docs) {
+        documents.push({
           ...(model.timestamps && {
             [getTimestampProperty('createdAt', model.timestamps)]: new Date(),
             [getTimestampProperty('updatedAt', model.timestamps)]: new Date(),
           }),
-          ...getDefaultValues(model.defaults),
+          ...(await getDefaultValues(model.defaults)),
           ...doc,
-        };
-      });
+        });
+      }
 
       const result = await model.collection.insertMany(
         documents as unknown as OptionalUnlessRequiredId<TSchema>[],
@@ -924,7 +928,7 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
           [getTimestampProperty('createdAt', model.timestamps)]: new Date(),
           [getTimestampProperty('updatedAt', model.timestamps)]: new Date(),
         }),
-        ...getDefaultValues(model.defaults),
+        ...(await getDefaultValues(model.defaults)),
         ...doc,
       };