--- a/src/model.ts
+++ b/src/model.ts
@@ -383,45 +383,46 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
         return;
       }
 
-      const finalOperations = [];
-      for (const op of operations) {
-        let operation = op;
-        if ('insertOne' in op) {
-          operation = {
-            insertOne: {
-              document: {
-                ...(await getDefaultValues(model.defaults)),
-                ...op.insertOne.document,
+      const finalOperations = await Promise.all(
+        operations.map(async (op) => {
+          let operation = op;
+          if ('insertOne' in op) {
+            operation = {
+              insertOne: {
+                document: {
+                  ...(await getDefaultValues(model.defaults)),
+                  ...op.insertOne.document,
+                },
               },
-            },
-          };
-        } else if (
-          'updateOne' in op &&
-          op.updateOne.upsert &&
-          !Array.isArray(op.updateOne.update)
-        ) {
-          const { update } = op.updateOne;
-          operation = {
-            updateOne: {
-              ...op.updateOne,
-              update: {
-                ...update,
-                $setOnInsert: cleanSetOnInsert(
-                  {
-                    ...(await getDefaultValues(model.defaults)),
-                    ...update.$setOnInsert,
-                  },
-                  update
-                ) as NonNullable<UpdateFilter<TSchema>['$setOnInsert']>,
+            };
+          } else if (
+            'updateOne' in op &&
+            op.updateOne.upsert &&
+            !Array.isArray(op.updateOne.update)
+          ) {
+            const { update } = op.updateOne;
+            operation = {
+              updateOne: {
+                ...op.updateOne,
+                update: {
+                  ...update,
+                  $setOnInsert: cleanSetOnInsert(
+                    {
+                      ...(await getDefaultValues(model.defaults)),
+                      ...update.$setOnInsert,
+                    },
+                    update
+                  ) as NonNullable<UpdateFilter<TSchema>['$setOnInsert']>,
+                },
               },
-            },
-          };
-        }
+            };
+          }
 
-        finalOperations.push(
-          model.timestamps ? timestampBulkWriteOperation(operation, model.timestamps) : operation
-        );
-      }
+          return model.timestamps
+            ? timestampBulkWriteOperation(operation, model.timestamps)
+            : operation;
+        })
+      );
 
       const result = await model.collection.bulkWrite(
         finalOperations as AnyBulkWriteOperation<TSchema>[],
@@ -872,17 +873,18 @@ export function build<TSchema extends BaseSchema, TOptions extends SchemaOptions
       docs: DocumentForInsert<TSchema, TOptions>[],
       options?: BulkWriteOptions
     ): Promise<TSchema[]> {
-      const documents = [];
-      for (const doc of docs) {
-        documents.push({
-          ...(model.timestamps && {
-            [getTimestampProperty('createdAt', model.timestamps)]: new Date(),
-            [getTimestampProperty('updatedAt', model.timestamps)]: new Date(),
-          }),
-          ...(await getDefaultValues(model.defaults)),
-          ...doc,
-        });
-      }
+      const documents = await Promise.all(
+        docs.map(async (doc) => {
+          return {
+            ...(model.timestamps && {
+              [getTimestampProperty('createdAt', model.timestamps)]: new Date(),
+              [getTimestampProperty('updatedAt', model.timestamps)]: new Date(),
+            }),
+            ...(await getDefaultValues(model.defaults)),
+            ...doc,
+          };
+        })
+      );
 
       const result = await model.collection.insertMany(
         documents as unknown as OptionalUnlessRequiredId<TSchema>[],