--- a/src/schema.ts
+++ b/src/schema.ts
@@ -15,7 +15,10 @@ export type SchemaTimestampOptions =
     }>
   | boolean;
 
-export type DefaultsOption<TProperties> = Partial<TProperties> | (() => Partial<TProperties>);
+export type DefaultsOption<TProperties> =
+  | Partial<TProperties>
+  | (() => Partial<TProperties>)
+  | (() => Promise<Partial<TProperties>>);
 
 export interface SchemaOptions<TProperties> {
   defaults?: DefaultsOption<TProperties>;