Hi @avaly,

Right now, we are using nanoid to generate some random tokens like:

```ts
import { schema, types } from 'papr';
import { nanoid } from 'nanoid'

const userSchema = schema({
  ...
  token: types.string({ required: true })
}, {
  defaults: () => ({
    token: nanoid();
  })
});

export type UserDocument = (typeof userSchema)[0];
export type UserOptions = (typeof userSchema)[1];
```

However, nanoid 4+ works only with ESM applications. Right now, we cannot migrated our app to ESM , but we've found the way to make it work with dynamic import:
```ts
import { schema, types } from 'papr';

async function getNanoId() {
  const nanoidImport = await import('nanoid');
  return nanoidImport.nanoid();
}

const userSchema = schema({
  ...
  token: types.string({ required: true })
}, {
  defaults: async () => ({
    token: await nanoid();
  })
});

export type UserDocument = (typeof userSchema)[0];
export type UserOptions = (typeof userSchema)[1];
```

I've updated the code to support async dynamic default values, so we can support this use case. It's the most straight forward way for us. Please let me know your thoughts on it.