@CarlosLozanoHealthCaters 2023-12-24
Sorry I couldn't run the benchmark in a CI, but I run it five times each in my computer and did an average. Here are my numbers:

Before:
- Object: 633,43 ops/s
- Function: 630,44 ops/s

After:
- Object: 624,45 ops/s
- Function: 651,85 ops/s
- Async: 643,13 ops/s

It's quite surprising that the new code is performing better (except direct object) and function and async is performing better than direct objects. 

If you are not really convinced on the implementation, I can propose another solution:
1. Add a new field to the schema options: defaultsType: 'static' (default) | 'dynamic' | 'async dynamic'
2. Depending on the defaultType, the code assign the defaults directly, or execute a function, or await an async function.
3. And the new option (defaultsType) can enforce the type of defaults field accordingly